Split up the flagging of and sending line to active dialogue views
Because the completion call of an active view can occur before other views are added to the set of active views this would result in the set being empty despite a bunch of other views still waiting for the line to come in.
The dialogue runner however sees an empty set of views and assumes all is good to continue.
These other views however would be off by one from the first view eventually resulting in an a null line being passed around and the view crashing.
Now all views are flagged as active before being given the line, preventing this from happening.

Fixes #97