Added a basic render detector to the sample scenes.
This detector is used to work out which render pipeline is in use.
The samples were all made using the built in render pipeline so when used in an URP/HDRP project The materials are all messed up.
This detector will fire off a warning that the materials will look weird until upgraded.
Not a huge fan of having a small object in each sample scene but the other solutions seem clunkier and this is very easy to delete.
In the long run would be neat to work out once the SRP upgrade has happened and then self-delete the object, but that is a future problem.

Fixes #98