removed auto continue as that is the default and views can handle it locally
This does mean there is one existing problem still to be solved, how to handle that interrupt and continue are 99% of the time going to be the same action/interaction as far as the user is concerned.
In the current system I have made a basic input view which sends over the interrupt signal but it has no knowledge of the state of other views so it won't know if it is continuing or interrupting.
May end up having to put the input handling side back into the line view.