default views now support nameplates.
Previously the character name was just a text object inside the LineView, so if you had multiple pieces of it they couldn't be enabled or disabled on the fly.
This adds a second container element for the line view to hold this nameplate aspect.
This behaviour has also been added to the OptionsListView