Automatic linking codegen can now be controlled via the YS settings.
Had to make a bunch of changes to a bunch of files and also add MiniJSON.
This is because the codegen pipeline can't use the Unity side of things so the built in JSON stuff is out.
Pretty much everything else is just some minor refactoring to make it less all jumbled together into a single file.