reworked the samples installer.
Now each approach for accessing the samples has their own file.
Currently that doesn't mean anything as three of them just open a url but in the future now all of that can be changed in the relevant place.
Also simplified the dialogue runner editor as a bonus of this.