Fix late line complete callback in line with markup
Fixes https://github.com/YarnSpinnerTool/YarnSpinner-Unity/issues/328

When showing characters with the typewriter effect the `LinePresenter`
loops the characters in the string and sets maxVisibleCharacters on
the Textmesh component. If a line has markup tags those will be looped
too. That means when a line contains markup it will exceed the number
of characters that needs to be passed to TM causing a delay between
the text becoming visible at runtime and the `OnLineDisplayComplete()`
method being called.

This can cause issues if, for example, a connected `ActionMarkupHandler`
needs to decide whether call `RequestHurryUpLine()` or `RequestNextLine()`
as the line will be visibly finished, but the loop will still be running
for the count of extra characters.

This replaces the total count in the loop with the amount of visible
characters so it will sync back up.

For some reason `lineText.textInfo` doesn't contain the correct current
count but calling `lineText.GetTextInfo()` with the current line text
does.