first step towards making the dialogue runner more async aware.
Currently it does this through a lot of completion sources, which isn't ideal.
A more complete solution is to make the VM itself async aware.
This is still WIP but impacts #311.

Not a full fix at this stage but now Stop() being async means it won't finish until AFTER on dialogue complete calls finish.
So while you can ignore the async nature of this method and call start this being odd is now more clearly signalled.