Undid changes of PR 335.
While the arguments of the original PR are valid, after discussion in the discord around this feature (https://discord.com/channels/754171172693868585/1072667932078903400/1426722073878724718) it was decided that the change isn't significant enough for breaking the existing codebases and flow. As a way to still get this behaviour we've added a Source field to LocalizedLine which will by default contain the dialogue runner.

Changes were manually undone due to other changes I wanted to persist that were in the same commit.