Bunch of about screen changes.
Redid it in UXML to hopefully make updating it easier.
Now has more links and useful info on it.
Now shows itself once on project launch.
The last shown version of the about page is now stored in the project settings.
The way the project was installed is now public.